test: cover image tool live providers
